+++ rpc/balance.go
@@ -21,6 +21,7 @@ type BalanceUpdate interface {
 type GenericBalanceUpdate struct {
 	Kind   string `json:"kind"`
 	Change int64  `json:"change,string"`
+	Origin string `json:"origin"` // block, migration
 }
 
 // BalanceUpdateKind returns the BalanceUpdateType's Kind field