+++ rpc/endorsement.go
@@ -11,8 +11,17 @@ import (
 // EndorsementOp represents an endorsement operation
 type EndorsementOp struct {
 	GenericOp
-	Level    int64                  `json:"level"`
-	Metadata *EndorsementOpMetadata `json:"metadata"`
+	Level       int64                  `json:"level"`       // <= v008
+	Metadata    *EndorsementOpMetadata `json:"metadata"`    // all protocols
+	Endorsement *EndorsementContent    `json:"endorsement"` // v009+
+	Slot        int                    `json:"slot"`        // v009+
+}
+
+func (e EndorsementOp) GetLevel() int64 {
+	if e.Endorsement != nil {
+		return e.Endorsement.Operations.Level
+	}
+	return e.Level
 }
 
 // EndorsementOpMetadata represents an endorsement operation metadata
@@ -24,4 +33,16 @@ type EndorsementOpMetadata struct {
 
 func (m EndorsementOpMetadata) Address() chain.Address {
 	return m.Delegate
+}
+
+// v009+
+type EndorsementContent struct {
+	Branch     string                `json:"branch"`
+	Operations EmbeddedEndorsementOp `json:"operations"`
+}
+
+// v009+
+type EmbeddedEndorsementOp struct {
+	Kind  chain.OpType `json:"kind"`
+	Level int64        `json:"level"`
 }